control set.seed() behavior
Changing the location of set.seed().
Currently it is set outside of the pop parameter loop. The sampling that happens inside that loop changes the seed slightly, so the order in which parameters are listed will affect the seed they get, altering the outcomes of the search. By setting it inside, each parameter should be starting with the same seed.